refactor(face-suggestions): move the suggestion engine into fork-owned files
person.service.ts carried the fork's entire face-suggestion engine inline —
the four suggestion-scan job handlers and the five confirm/reject/ignore/
dismiss/get endpoints — none of which interleave with upstream PersonService
logic. Moves that code, verbatim, into a new FaceSuggestionService and
FaceSuggestionController, matching the isolation already applied to
face-repair.service.ts, classification.service.ts and shared-space.service.ts.
person.controller.ts returns to its origin/main content; person.service.ts
keeps only the genuine in-place hooks (verdict clearing in reassignFaces/
reassignFacesById, the re-scan queue in update(), the backfill-completion
queue, and the bootstrap sweep).

Medium tests that exercised the moved methods through PersonService now use
a second service instance sharing the same MediumTestContext dependencies
(MediumTestContext.getService), added to test/medium.factory.ts for this
purpose.

Behaviour is unchanged: every route path, DTO and job name is identical, the
unit suite's pass count is identical before and after (5655 passed / 14
skipped), and rebuilding + regenerating the OpenAPI spec and both generated
clients produces a zero diff in open-api/, packages/sdk/ and mobile/openapi/.
