+++ server/src/controllers/index.ts
@@ -11,6 +11,7 @@ import { DatabaseBackupController } from 'src/controllers/database-backup.contro
 import { DownloadController } from 'src/controllers/download.controller';
 import { DuplicateController } from 'src/controllers/duplicate.controller';
 import { FaceRepairAdminController } from 'src/controllers/face-repair-admin.controller';
+import { FaceSuggestionController } from 'src/controllers/face-suggestion.controller';
 import { FaceController } from 'src/controllers/face.controller';
 import { GalleryMapController } from 'src/controllers/gallery-map.controller';
 import { IntegrityAdminController } from 'src/controllers/integrity-admin.controller';
@@ -62,6 +63,7 @@ export const controllers = [
   DuplicateController,
   FaceController,
   FaceRepairAdminController,
+  FaceSuggestionController,
   GalleryMapController,
   IntegrityAdminController,
   JobController,
