+++ server/src/controllers/person.controller.ts
@@ -21,7 +21,6 @@ import { AuthDto } from 'src/dtos/auth.dto';
 import {
   AssetFaceUpdateDto,
   DetachScopedPersonDto,
-  FaceSuggestionActionResponseDto,
   MergePersonDto,
   MergeScopedPeopleDto,
   PeopleFaceStatisticsResponseDto,
@@ -31,9 +30,6 @@ import {
   PersonCreateDto,
   PersonFacePageQueryDto,
   PersonFacePageResponseDto,
-  PersonFaceSuggestionPageQueryDto,
-  PersonFaceSuggestionPageResponseDto,
-  PersonFaceSuggestionParamsDto,
   PersonResponseDto,
   PersonSearchDto,
   PersonStatisticsResponseDto,
@@ -309,103 +305,5 @@ export class PersonController {
     @Body() dto: MergePersonDto,
   ): Promise<BulkIdResponseDto[]> {
     return this.service.mergePerson(auth, id, dto);
-  }
-
-  // F21: publishes the permission getFaceSuggestions actually enforces. Deliberately PersonUpdate, not
-  // PersonRead — PersonRead also resolves via access.person.checkSharedSpaceAccess (see
-  // src/utils/access.ts), which would let a space member read the owner's whole-library pending review
-  // queue (D6, see the comment on getFaceSuggestions in person.service.ts). Do not relax this back to
-  // PersonRead to "match" a shared-space caller; the service enforcement is the source of truth here.
-  @Get(':id/face-suggestions')
-  @Authenticated({ permission: Permission.PersonUpdate })
-  @Endpoint({
-    summary: 'Get face suggestions for a person',
-    description: 'Retrieve near-miss unassigned faces suggested for this person, best match first.',
-    history: new HistoryBuilder().added('v1').beta('v1'),
-  })
-  getPersonFaceSuggestions(
-    @Auth() auth: AuthDto,
-    @Param() { id }: UUIDParamDto,
-    @Query() dto: PersonFaceSuggestionPageQueryDto,
-  ): Promise<PersonFaceSuggestionPageResponseDto> {
-    return this.service.getFaceSuggestions(auth, id, dto);
-  }
-
-  // F21: publishes PersonUpdate, the person-level permission confirmFaceSuggestion enforces — not
-  // PersonReassign, which the service never checks. confirmFaceSuggestion ALSO enforces PersonCreate on
-  // the face itself (assetFaceId), but the guard can only carry one permission; that face-level check
-  // stays service-level (see the comment on confirmFaceSuggestion in person.service.ts). Do not drop it
-  // there on the assumption this decorator covers it.
-  //
-  // S11 (F24): the response EXPLICITLY reports whether the call acted or was a no-op — the service's return
-  // value, not a fixed default. The web modal used to infer "already resolved" from a 400, which is
-  // indistinguishable from a genuine authorization failure (see the comment that used to sit here and on
-  // PersonSuggestionReviewModal.svelte).
-  //
-  // S11b (F24): that report is the `acted` field of the BODY, always under 200 — NOT a 200-vs-204 status
-  // code. @oazapfts/runtime's ok() resolves to the body and throws away the numeric status for every
-  // success code, so no generated client can read a status-code signal. Do not "simplify" this back to
-  // @HttpCode + res.status(): it compiles, tests green against supertest, and is unusable from the SDK.
-  @Post(':id/face-suggestions/:assetFaceId/confirm')
-  @Authenticated({ permission: Permission.PersonUpdate })
-  @HttpCode(HttpStatus.OK)
-  @Endpoint({
-    summary: 'Confirm a face suggestion',
-    description: 'Assign the suggested face to the person. Idempotent — the response reports whether it acted.',
-    history: new HistoryBuilder().added('v1').beta('v1'),
-  })
-  async confirmPersonFaceSuggestion(
-    @Auth() auth: AuthDto,
-    @Param() { id, assetFaceId }: PersonFaceSuggestionParamsDto,
-  ): Promise<FaceSuggestionActionResponseDto> {
-    return { acted: await this.service.confirmFaceSuggestion(auth, id, assetFaceId) };
-  }
-
-  @Post(':id/face-suggestions/:assetFaceId/reject')
-  @Authenticated({ permission: Permission.PersonUpdate })
-  @HttpCode(HttpStatus.OK)
-  @Endpoint({
-    summary: 'Reject a face suggestion',
-    description:
-      'Reject this suggestion for the person. The face stays unassigned. Idempotent — the response reports whether it acted.',
-    history: new HistoryBuilder().added('v1').beta('v1'),
-  })
-  async rejectPersonFaceSuggestion(
-    @Auth() auth: AuthDto,
-    @Param() { id, assetFaceId }: PersonFaceSuggestionParamsDto,
-  ): Promise<FaceSuggestionActionResponseDto> {
-    return { acted: await this.service.rejectFaceSuggestion(auth, id, assetFaceId) };
-  }
-
-  @Post(':id/face-suggestions/:assetFaceId/ignore')
-  @Authenticated({ permission: Permission.PersonUpdate })
-  @HttpCode(HttpStatus.OK)
-  @Endpoint({
-    summary: 'Ignore a face suggestion',
-    description:
-      'Ignore this suggestion for the person. The face stays unassigned. Idempotent — the response reports whether it acted.',
-    history: new HistoryBuilder().added('v1').beta('v1'),
-  })
-  async ignorePersonFaceSuggestion(
-    @Auth() auth: AuthDto,
-    @Param() { id, assetFaceId }: PersonFaceSuggestionParamsDto,
-  ): Promise<FaceSuggestionActionResponseDto> {
-    return { acted: await this.service.ignoreFaceSuggestion(auth, id, assetFaceId) };
-  }
-
-  @Post(':id/face-suggestions/:assetFaceId/dismiss')
-  @Authenticated({ permission: Permission.PersonUpdate })
-  @HttpCode(HttpStatus.OK)
-  @Endpoint({
-    summary: 'Dismiss a face suggestion',
-    description:
-      'Compatibility alias for rejecting this suggestion. The face stays unassigned. Idempotent — the response reports whether it acted.',
-    history: new HistoryBuilder().added('v1').beta('v1'),
-  })
-  async dismissPersonFaceSuggestion(
-    @Auth() auth: AuthDto,
-    @Param() { id, assetFaceId }: PersonFaceSuggestionParamsDto,
-  ): Promise<FaceSuggestionActionResponseDto> {
-    return { acted: await this.service.dismissFaceSuggestion(auth, id, assetFaceId) };
   }
 }
