+++ server/src/services/index.ts
@@ -14,6 +14,7 @@ import { DatabaseService } from 'src/services/database.service';
 import { DownloadService } from 'src/services/download.service';
 import { DuplicateService } from 'src/services/duplicate.service';
 import { FaceRepairService } from 'src/services/face-repair.service';
+import { FaceSuggestionService } from 'src/services/face-suggestion.service';
 import { HlsService } from 'src/services/hls.service';
 import { IntegrityService } from 'src/services/integrity.service';
 import { JobService } from 'src/services/job.service';
@@ -75,6 +76,7 @@ export const services = [
   DownloadService,
   DuplicateService,
   FaceRepairService,
+  FaceSuggestionService,
   IntegrityService,
   HlsService,
   JobService,
